refactor: centralize service and repository
