+++ src/main/java/org/storkforge/barkr/api/controller/ApiAccountController.java
@@ -4,7 +4,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.storkforge.barkr.domain.services.api.ApiAccountService;
+import org.storkforge.barkr.domain.AccountService;
 import org.storkforge.barkr.dto.accountDto.ResponseAccount;
 import org.storkforge.barkr.dto.accountDto.ResponseAccountList;
 
@@ -12,9 +12,9 @@ import org.storkforge.barkr.dto.accountDto.ResponseAccountList;
 @RequestMapping("/api")
 public class ApiAccountController {
 
-    private final ApiAccountService accountService;
+    private final AccountService accountService;
 
-    public ApiAccountController(ApiAccountService accountService) {
+    public ApiAccountController(AccountService accountService) {
         this.accountService = accountService;
     }
 
@@ -25,6 +25,6 @@ public class ApiAccountController {
 
     @GetMapping("/accounts/{id}")
     public ResponseAccount account(@PathVariable Long id) {
-        return accountService.findOne(id);
+        return accountService.findById(id);
     }
 }
