+++ src/main/java/org/storkforge/barkr/exceptions/PostNotFound.java
@@ -6,7 +6,7 @@ import org.storkforge.barkr.domain.entity.Post;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class PostNotFound extends RuntimeException {
-    public PostNotFound(Post post) {
+    public PostNotFound() {
         super();
     }
