+++ src/main/java/org/storkforge/barkr/graphql/AccountResolver.java
@@ -1,13 +1,11 @@
-package org.storkforge.barkr.web.graphql;
-
-
+package org.storkforge.barkr.graphql;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.stereotype.Controller;
-import org.storkforge.barkr.web.domain.AccountService;
+import org.storkforge.barkr.domain.AccountService;
 import org.storkforge.barkr.dto.accountDto.ResponseAccount;
 
 import java.util.List;
