+++ src/main/java/org/storkforge/barkr/graphql/PostResolver.java
@@ -1,11 +1,11 @@
-package org.storkforge.barkr.web.graphql;
+package org.storkforge.barkr.graphql;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.stereotype.Controller;
-import org.storkforge.barkr.web.domain.PostService;
+import org.storkforge.barkr.domain.PostService;
 import org.storkforge.barkr.dto.postDto.ResponsePost;
 
 import java.util.List;
@@ -31,7 +31,7 @@ public class PostResolver {
     @QueryMapping("post")
     public ResponsePost post(@Argument @NotNull @Positive Long id) {
         try {
-            ResponsePost post = postService.findOne(id);
+            ResponsePost post = postService.findById(id);
             if (post == null) {
                 throw new RuntimeException("Post not found for ID: " + id);
             }
