+++ src/main/java/org/storkforge/barkr/infrastructure/persistence/AccountRepository.java
@@ -1,4 +1,4 @@
-package org.storkforge.barkr.web.infrastructure.persistence;
+package org.storkforge.barkr.infrastructure.persistence;
 
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.ListCrudRepository;
@@ -10,5 +10,5 @@ public interface AccountRepository extends ListCrudRepository<Account, Long> {
   Optional<Account> findByUsernameEqualsIgnoreCase(String username);
 
   @Query("SELECT a.image FROM Account a WHERE a.id = :id")
-  byte[] getAccountImage(Long id);
+  Optional<byte[]> getAccountImage(Long id);
 }
