+++ src/main/java/org/storkforge/barkr/web/controller/WebController.java
@@ -1,5 +1,8 @@
 package org.storkforge.barkr.web.controller;
 
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
@@ -9,9 +12,9 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.context.request.async.DeferredResult;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import org.storkforge.barkr.web.domain.AccountService;
-import org.storkforge.barkr.web.domain.DogFactService;
-import org.storkforge.barkr.web.domain.PostService;
+import org.storkforge.barkr.domain.AccountService;
+import org.storkforge.barkr.domain.DogFactService;
+import org.storkforge.barkr.domain.PostService;
 import org.storkforge.barkr.dto.accountDto.ResponseAccount;
 import org.storkforge.barkr.dto.postDto.CreatePost;
 
@@ -56,7 +59,7 @@ public class WebController {
   }
 
   @GetMapping("/{username}")
-  public DeferredResult<String> user(@PathVariable("username") String username, Model model, RedirectAttributes redirectAttributes) {
+  public DeferredResult<String> user(@PathVariable("username") @NotBlank String username, Model model, RedirectAttributes redirectAttributes) {
     DeferredResult<String> result = new DeferredResult<>();
     ResponseAccount queryAccount;
 
@@ -89,7 +92,7 @@ public class WebController {
   }
 
   @PostMapping("/post/add")
-  public String addPost(@ModelAttribute CreatePost dto,
+  public String addPost(@ModelAttribute @NotNull CreatePost dto,
                         RedirectAttributes redirectAttributes) {
     if (dto.content() == null || dto.content().trim().isEmpty()) {
       redirectAttributes.addFlashAttribute("error", "Post content cannot be empty");
@@ -111,7 +114,7 @@ public class WebController {
   }
 
   @GetMapping("/account/{id}/image")
-  public ResponseEntity<byte[]> getAccountImage(@PathVariable("id") Long id) throws IOException {
+  public ResponseEntity<byte[]> getAccountImage(@PathVariable("id") @Positive @NotNull Long id) throws IOException {
     byte[] accountImage = accountService.getAccountImage(id);
 
     if (accountImage == null) {
@@ -127,7 +130,7 @@ public class WebController {
   }
 
   @PostMapping("/account/{id}/upload")
-  public String uploadImage(@PathVariable Long id, @RequestParam("file") MultipartFile file, RedirectAttributes redirectAttributes) throws IOException {
+  public String uploadImage(@PathVariable @Positive @NotNull Long id, @RequestParam("file") @NotNull MultipartFile file, RedirectAttributes redirectAttributes) throws IOException {
     if (file.isEmpty()) {
       redirectAttributes.addFlashAttribute("error", "File is empty");
       return "redirect:/";
